Qcamera2: HAL: Optimising buffer count for Blob.
Issue:
When the snapshot size is 21MP, HAL is allocating 19
buffers (10 for snapshot and 9 for reprocess) of ~32MB
size and frameworks is also allocating the buffers of
same size. In some targets free RAM is less, therefore
allocating these many buffers of ~32MB size will trigger
lowmemorykiller and cause the cts process to kill.

Fix:
Reduced the number of buffers to 6 for Blob as it is
not affecting any usecase and performance.

Change-Id: If6fe80b67ebc04b8dc01e559cd1f561805afb63b